Group the refresh and bookmark buttons from the location bar into a `location-buttons` component
