+++ app/ui/browser-modern/views/browser/location/index.jsx
@@ -15,9 +15,9 @@ import PureRenderMixin from 'react-addons-pure-render-mixin';
 import { connect } from 'react-redux';
 
 import Style from '../../../../shared/style';
-import Btn from '../../../../shared/widgets/btn';
 import VerticalSeparator from '../../../../shared/widgets/vertical-separator';
 import LocationInfo from './location-info';
+import LocationButtons from './location-buttons';
 import AutocompletedSearch from '../../../../shared/widgets/autocompleted-search';
 import AutocompletionListItem from './autocompletion-list-item';
 
@@ -26,14 +26,13 @@ import * as SharedPropTypes from '../../../model/shared-prop-types';
 import * as LocationUtil from '../../../../shared/util/location-util';
 import * as UISelectors from '../../../selectors/ui';
 import * as PagesSelectors from '../../../selectors/pages';
-import * as ProfileSelectors from '../../../selectors/profile';
 import * as ProfileEffects from '../../../actions/profile-effects';
 import * as PageEffects from '../../../actions/page-effects';
 
 const LOCATION_BAR_STYLE = Style.registerStyle({
   flex: 1,
   alignSelf: 'stretch',
-  alignItems: 'center',
+  alignItems: 'stretch',
   margin: '8px 0px',
   padding: '0 5px',
   borderRadius: 'var(--theme-default-roundness)',
@@ -43,17 +42,12 @@ const LOCATION_BAR_STYLE = Style.registerStyle({
 
 const LOCATION_BAR_INPUT_STYLE = Style.registerStyle({
   flex: 1,
-  alignSelf: 'stretch',
   padding: '0 1px',
   fontSize: '110%',
 });
 
 const DROPDOWN_LIST_STYLE = Style.registerStyle({
   maxHeight: `calc(100vh - ${NAVBAR_HEIGHT}px - ${TABBAR_HEIGHT}px)`,
-});
-
-const LOCATION_BAR_REFRESH_BUTTON_STYLE = Style.registerStyle({
-  padding: '2px',
 });
 
 const SEPARATOR_STYLE = Style.registerStyle({
@@ -83,16 +77,6 @@ class Location extends Component {
     this.props.dispatch(PageEffects.navigatePageTo(this.props.pageId, location));
   }
 
-  handleBookmarkButtonClick = () => {
-    const pageId = this.props.pageId;
-    const bookmarked = this.props.pageIsBookmarked;
-    this.props.dispatch(ProfileEffects.setRemoteBookmarkState(pageId, !bookmarked));
-  }
-
-  handleRefreshButtonClick = () => {
-    this.props.dispatch(PageEffects.navigatePageRefresh(this.props.pageId));
-  }
-
   render() {
     return (
       <div className={`browser-location ${LOCATION_BAR_STYLE}`}>
@@ -107,27 +91,7 @@ class Location extends Component {
           dataSrc={this.props.locationAutocompletions}
           childComponent={AutocompletionListItem}
           dropdownListClassName={DROPDOWN_LIST_STYLE} />
-        <Btn title="Bookmark"
-          width="18px"
-          height="18px"
-          image="toolbar.png"
-          imgWidth="792px"
-          imgHeight="72px"
-          imgPosition={`${this.props.pageIsBookmarked ? '-144px' : '-126px'} 0px`}
-          imgPositionHover={`${this.props.pageIsBookmarked ? '-144px' : '-126px'} -18px`}
-          onClick={this.handleBookmarkButtonClick} />
-        <VerticalSeparator className={SEPARATOR_STYLE} />
-        <Btn title="Refresh"
-          className={LOCATION_BAR_REFRESH_BUTTON_STYLE}
-          width="14px"
-          height="14px"
-          image="reload-stop-go.png"
-          imgWidth="42px"
-          imgHeight="28px"
-          imgPosition="0px 0px"
-          imgPositionActive="0px -14px"
-          disabled={!this.props.pageCanRefresh}
-          onClick={this.handleRefreshButtonClick} />
+        <LocationButtons pageId={this.props.pageId} />
       </div>
     );
   }
@@ -139,8 +103,6 @@ Location.propTypes = {
   dispatch: PropTypes.func.isRequired,
   pageId: PropTypes.string.isRequired,
   pageLocation: PropTypes.string.isRequired,
-  pageCanRefresh: PropTypes.bool.isRequired,
-  pageIsBookmarked: PropTypes.bool.isRequired,
   locationAutocompletions: SharedPropTypes.LocationAutocompletions,
 };
 
@@ -148,8 +110,6 @@ function mapStateToProps(state, ownProps) {
   const page = PagesSelectors.getPageById(state, ownProps.pageId);
   return {
     pageLocation: page ? page.location : '',
-    pageCanRefresh: page ? PagesSelectors.getPageCanRefresh(state, ownProps.pageId) : false,
-    pageIsBookmarked: page ? ProfileSelectors.isBookmarked(state, page.location) : false,
     locationAutocompletions: page ? UISelectors.getLocationAutocompletions(state, page.id) : null,
   };
 }
